TensorFlow/contrib/cv/AnimeFaceGAN_ID1062_for_Tensorflow/requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
cycler==0.10.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.35.0
h5py==3.1.0
imageio==2.16.2
importlib-metadata==3.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
numpy==1.20.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
Pillow==9.1.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.7.3
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tqdm==4.56.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
